requirements.txt
newspaper3k
langchain
langchain-google-genai
streamlit
python-dotenv
beautifulsoup4
datetime
pandas
langchain-openai
cloudinary
flask
flask-cors
docx